Seperated input and output rsyslog configurations
Seperated out the viaq and example tasks and defaults from
the main rsyslog tasks and defaults and started initial
seperation between input and output configurations.

Added sub roles to deploy the different configuration
files based on the input logs collections and the output
they need to be sent to.

Created initial role for oVirt configurations.

Signed-off-by: Shirly Radco <sradco@redhat.com>