Make logging role idempotent
We set the state of the rsyslog systemd service to "restarted" when
we enable the service, which will always trigger a restart.  We already
use a handler to restart the rsyslog service when we modify the config.
We can safely set the state to "started" when we enable the rsyslog
service, which will not trigger a change if the service is already
enabled and running.