RHELPLAN-33869 - Changing the default value of rsyslog_default from true to false
If rsyslog_default is set to true, it deploys the default rsyslog.conf to /etc.
The rsyslog.conf contains the initial input and output plugins for the system logs.
The logging role adds the functionality to process logs including the system logs.
For instance, which input roles to be used, which output roles to be used, how the
log messages are normalized and distributed.  The default rsyslog.conf could conflict
with the functionality. There is one particular use case which needs rsyslog_default
set to true.  Ovirt already relies on the default rsyslog.conf for the system log
handling. To guarantee it does not break the ovirt logging deployment, we chose to
keep the rsyslog_default var and ability to deploy the default rsyslog.conf for now.

Additionally, Changing logs_collection parameter(s) as follows:
[current]
logging_outputs:
  -name: <output_role_name>
   type: <output_type>
   logs_collections:
     - name: <input_role_name>  # type of the input_role
[new]
logging_outputs:
  -name: <output_role_name>
   type: <output_type>
   logs_collections:
     - name: <input_role_name>  # unique name
       type: <input_role_type>  # type of the input_role

where <input_role_name> in the "current" format specifies the type of
the input_role in the sense of the output_role convention.  To reduce
the confusion, it'd be preferable that "name" and "type" have the same
meaning between output_role (logging_outputs) and input_role (logs_collections).