do not use ../ in path when including roles
For some reason, when including parent roles from a sub-role
like this:
```yaml
  include_role:
    name: "{{ role_path }}/../../../"
    tasks_from: deploy.yml
```
and there are handlers to be included and executed from the included
role, ansible thinks the handler tasks are separate, and executes
them separately, leading to a problem that the same handler is
executed multiple times in a row.  Using the role path without
`../` allows ansible to correctly determine that the included
handlers are all the same task, and only executes them once.