Fixing the key/cert setting parameter use_cert for the elasticsearch output
There used to be a global variable use_local_omelasticsearch_cert to indicate
indicate key/certificates on the local host to be copied to the target host.
In this commit, the global variable is eliminated. To specify the local key/
certificate files to be copied, just set the local paths to key_src, cert_src,
and ca_cert_src.