Support imuxsock in the basics input.
RHEL-7 systemd contains listen.conf in the package to be placed in /etc/rsyslog.d,
which expects the imuxsock configuration in the old style. (See [1] for the details.)
This PR let RHEL-7 use the old rsyslog syntax style to handle imuxsock and the rest
platforms do the new style.

[1] - https://github.com/linux-system-roles/logging/pull/124#issuecomment-636193575

CI-tests
- adding tests_imuxsock_files.yml
- checking whether "a RainerScript command" string is in the systemctl status output since
  when the conflict caused by listen.conf occurs, a message "command 'SystemLogSocketName'
  is currently not permitted - did you already set it via a RainerScript command" is issued
  by rsyslogd.