support jinja 2.7
Do not use the case_sensitive argument to the unique filter
For doing selectattr with string comparison, use `match` with
an argument like `^STRING$` instead of `==` or `equalto` with
`STRING`.
For doing selectattr with boolean values, you don't need to
provide a test - `selectattr('boolvar')` is the same as
`selectattr('boolvar', '==', true)`.  For `false` comparison,
use `rejectattr('boolvar')` which is the same as
selectattr('boolvar', '==', false)`
Do not use `namespace` as it is only needed with loops, not `if`
Also includes support for ansible 2.8 and/or python2, which when
running in certain environments converts the strings to unicode
like `u'val'`.