Issue: cert and key files should be deployed in the /etc/pki/tls folder
When ca_cert_src, cert_src, and private_key_src are specified without the
corresponding target location - ca_cert, cert, and private_key, the files
are copied to /etc/rsyslog.d for the elasticsearch output and the relp
input/output. On the other hand, logging_pki_files copies to /etc/pki/tls.

The inconsistent behaviour would confuse the users. Since relp is a new
feature, and we are releasing the next version as a major update (with
major_version++), it is a good timing to adjust it.

README and test cases are updated.

https://github.com/linux-system-roles/logging/issues/205