RHELPLAN-67347 - Logging - Integrating ELK with RHV-4.4 fails as RHVH is missing 'rsyslog-gnutls' package.
rsyslog-gnutls is needed when the global tls is configured with the
variable logging_pki_files. When the package module is called for
rsyslog-gnutls, this pr adds a check for logging_pki-files and only
when it is configured, rsyslog-gnutls is installed.

Also, a check is added to every test case that the necessary and
sufficient packages are installed.

With bug fixes found in the reviews by @richm. Thanks!

https://bugzilla.redhat.com/show_bug.cgi?id=1927943