Add user and password in elasticsearch output
Bug-Url: https://bugzilla.redhat.com/1990490
Signed-off-by: Aviv Litman <alitman@redhat.com>

Update output_elasticsearch.j2

Remove the 'tls' from the 'uid' and 'pwd'  parameters.

Update output_elasticsearch.j2

The tests_ovirt_elasticsearch.yml test fails, therefore fix it.

- 31-output-elasticsearch-elasticsearch_output*.conf: set mode to 0600.
- output_elasticsearch.j2: eliminate unnecessary __uid and __pwd
  setting.

- Introducing logging_elasticsearch_password and deploy_nolog.yml.
- Adding the params to tests_ovirt_elasticsearch.yml.

uid and pwd should be available outside of tls

uid should only be defined if password is defined

uid should only be defined if password is defined and not empty.

Unless both uid value in the elasticsearch output and the global var
logging_elasticsearch_password are configured, the pair uid and pwd
will not be put in the rsyslog omelasticsearch config file.

- The elasticsearch output calls no_log version of deploy.yml only
  if the user has provided a password.
- Add the uid and logging_elasticsearch_password to README.md.
- Add the test case to tests_ovirt_elasticsearch.yml.