make purge idempotent; get rid of reset
The use of `logging_purge_confs` should be be idempotent.  This is done
by first getting a list of all unowned config files to see if there are
any.  If there are, then the role will remove all unowned config files
not being currently written by the role invocation.  That way, if you
run the role again with `logging_purge_confs: true` and the same
parameters, the role will detect that there should be no changes to the
existing files, and return `changed: false`.

Get rid of `logging_reset_confs` - use only purge.  The way purge
works now is that it will look if the rsyslog package needs to be
reinstalled, by checking if /etc/rsyslog.conf is modified, and the
user has not provided any rsyslog_inputs (because rsyslog_inputs
will trigger /etc/rsyslog.conf to be overwritten if __rsyslog_enabled).