ci: Avoid variable name conflicts
If the same variable name is used in the `vars` declaration
and the target of `set_fact` in the separate test playbooks,
the value set in `set_fact` overwrites the value declared in
`vars` and it makes the latter test fail when you execute the
two tests in one ansible-playbook.

This patch replaces the variable names used for `set_fact` not
to conflict with the tests declaring in `vars`.

Signed-off-by: Noriko Hosoi <nhosoi@icloud.com>