fix: ansible-core-2.16 - only use to_nice_json for output formatting
Cause: The code was using `to_nice_json` to construct a list used for
further processing.

Consequence: The variable was evaluated as a `string`, not a `list`
and was causing no files to match.

Fix: Remove the `to_nice_json` so the variable is evaluated as
a `list`.

Result: The code works as expected.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>