fix: Remove packages with command line
Cause: Fedora ≥ 411 use dnf5 now, and there Ansible's `package:` does not
automatically remove reverse dependencies such as `rsyslog-gnutls` any
more. That's a good thing in general to avoid damage/surprised; but in
our role, the definition, documentation, and tests rely on the reverse
dependency removal for proper cleanup/reset.

Consequence: The role failed when trying to reset the configuration if
there are any rsyslog-* extensions installed.

Fix: Do the package removal with calling yum/dnf directly.

Result: The role works on Fedora ≥ 41 as well.

Note that we can drop the `use:`. At this point we already checked
__logging_is_ostree in the parent block.

Signed-off-by: Martin Pitt <mpitt@redhat.com>