fix: Skip runtime operations in non-systemd environments
Cause: Both the actual role and the tests were previously assuming that
the system was booted with systemd and could run services. But this is
not the case when running the role during container builds.

Consequence: The role did not work during bootc container builds.

Fix: Detect if the system is booted (with systemd), and skip all runtime
operations and checks if not. Add clear error messages if systemd is not
installed, or logging_certificates is set during a container build; this
role is not supported then.

Result: The role now works during container builds.

Skip tests_server.yml for the time being. This uses the certificate
role, which cannot currently work during a container build. We need to
provide a solution for this at some point, but for now skip that use
case and fail with a "not supported" error.

https://issues.redhat.com/browse/RHEL-88689